// lib/types.ts
export interface ElectoralData {
  distritoFederal: string
  distritoFederalNombre: string
  distritoLocal: string
  distritoLocalNombre: string
  municipio: string
  municipioNombre: string
  localidad: string
  localidadNombre: string
  seccion: string
  tipo: string
  ubicacion: string
  presidente: string
  secretario1: string
  secretario2: string
  escrutador1: string
  escrutador2: string
  escrutador3: string
  suplente1: string
  suplente2: string
  suplente3: string
}